requirements.txt
python-dotenv
groq
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
oauth2client
